Validate IncomeBenefit to ensure compliance
